refactor: change components for country autocomlete component
