--- src/components/CountryAutocomplete/index.ts
@@ -1,1 +1,0 @@
-export { default as CountryAutocomplete } from './CountryAutocomplete';
