+++ src/components/CountryAutocompleteControl/index.ts
@@ -1,0 +1,1 @@
+export { default as CountryAutocompleteControl } from './CountryAutocompleteControl';
