+++ src/components/CountryAutocompleteNoControl/CountryAutocompleteNoControl.module.scss
@@ -1,0 +1,3 @@
+.root {
+  position: relative;
+}
