+++ src/components/CountryList/CountryList.module.scss
@@ -1,8 +1,4 @@
 .root {
-  position: relative;
-}
-
-.list {
   position: absolute;
   top: 100%;
   left: 0;
