+++ src/components/CountryList/index.ts
@@ -1,0 +1,1 @@
+export { default as CountryList } from './CountryList';
